# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UK Support",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/staff/tickets": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "List Tickets",
        "operationId": "list_tickets_api_v1_staff_tickets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "building_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 9223372036854775807,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Building Id"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 9223372036854775807,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "mine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Mine"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TicketListItemResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/buildings": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "List Buildings",
        "operationId": "list_buildings_api_v1_staff_buildings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BuildingShortResponse"
                  },
                  "type": "array",
                  "title": "Response List Buildings Api V1 Staff Buildings Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/staff/categories": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "List Categories",
        "operationId": "list_categories_api_v1_staff_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryShortResponse"
                  },
                  "type": "array",
                  "title": "Response List Categories Api V1 Staff Categories Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/staff/tickets/unread-count": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "Unread Count",
        "operationId": "unread_count_api_v1_staff_tickets_unread_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/staff/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "Get Ticket",
        "operationId": "get_ticket_api_v1_staff_tickets__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/tickets/{ticket_id}/status": {
      "post": {
        "tags": [
          "staff"
        ],
        "summary": "Change Status",
        "operationId": "change_status_api_v1_staff_tickets__ticket_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/tickets/{ticket_id}/messages": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_api_v1_staff_tickets__ticket_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response List Messages Api V1 Staff Tickets  Ticket Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "staff"
        ],
        "summary": "Send Message",
        "operationId": "send_message_api_v1_staff_tickets__ticket_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_message_api_v1_staff_tickets__ticket_id__messages_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/tickets/{ticket_id}/read": {
      "post": {
        "tags": [
          "staff"
        ],
        "summary": "Mark Read",
        "operationId": "mark_read_api_v1_staff_tickets__ticket_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_send_message_api_v1_staff_tickets__ticket_id__messages_post": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          }
        },
        "type": "object",
        "title": "Body_send_message_api_v1_staff_tickets__ticket_id__messages_post"
      },
...
      "BuildingShortResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "address"
        ],
        "title": "BuildingShortResponse"
      },
...
      "CategoryShortResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "CategoryShortResponse"
      },
...
      "FileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "original_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Name"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "mime",
          "size",
          "original_name",
          "url"
        ],
        "title": "FileResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "sender_type": {
            "$ref": "#/components/schemas/SenderType"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserShortResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileResponse"
            },
            "type": "array",
            "title": "Files"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticket_id",
          "sender_type",
          "author",
          "text",
          "files",
          "created_at"
        ],
        "title": "MessageResponse"
      },
...
      "PaginatedResponse_TicketListItemResponse_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TicketListItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "PaginatedResponse[TicketListItemResponse]"
      },
...
      "SenderType": {
        "type": "string",
        "enum": [
          "CLIENT",
          "STAFF",
          "SYSTEM"
        ],
        "title": "SenderType"
      },
...
      "StatusChangeRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusChangeRequest"
      },
...
      "TicketDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/TicketType"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryShortResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "building": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuildingShortResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "apartment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apartment"
          },
          "client": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserShortResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_client_message_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Client Message At"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "preferred_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Time"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileResponse"
            },
            "type": "array",
            "title": "Files"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/StatusChangeResponse"
            },
            "type": "array",
            "title": "History"
          },
          "allowed_statuses": {
            "items": {
              "$ref": "#/components/schemas/TicketStatus"
            },
            "type": "array",
            "title": "Allowed Statuses"
          },
          "unread": {
            "type": "boolean",
            "title": "Unread",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "priority",
          "description",
          "category",
          "building",
          "apartment",
          "client",
          "assignee",
          "created_at",
          "last_client_message_at",
          "contact_phone",
          "preferred_time",
          "rating",
          "closed_at",
          "files",
          "history",
          "allowed_statuses",
          "unread"
        ],
        "title": "TicketDetailResponse"
      },
      "TicketListItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/TicketType"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryShortResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "building": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuildingShortResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "apartment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apartment"
          },
          "client": {
            "$ref": "#/components/schemas/UserShortResponse"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserShortResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_client_message_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Client Message At"
          },
          "unread": {
            "type": "boolean",
            "title": "Unread",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "priority",
          "description",
          "category",
          "building",
          "apartment",
          "client",
          "assignee",
          "created_at",
          "last_client_message_at",
          "unread"
        ],
        "title": "TicketListItemResponse"
      },
      "TicketPriority": {
        "type": "string",
        "enum": [
          "NORMAL",
          "URGENT"
        ],
        "title": "TicketPriority"
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "NEW",
          "IN_PROGRESS",
          "WAITING_CLIENT",
          "CLOSED",
          "REJECTED"
        ],
        "title": "TicketStatus"
      },
      "TicketType": {
        "type": "string",
        "enum": [
          "REQUEST",
          "QUESTION"
        ],
        "title": "TicketType"
      },
...
      "UnreadCountResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "UnreadCountResponse"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "max_user_id": {
            "type": "integer",
            "title": "Max User Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "max_user_id",
          "first_name",
          "last_name",
          "username",
          "phone",
          "role"
        ],
        "title": "UserResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "CLIENT",
          "MANAGER",
          "ADMIN"
        ],
        "title": "UserRole"
      },
      "UserShortResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name"
        ],
        "title": "UserShortResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}